Added configuring ascii art placement and alignment
You can now place your ascii art to the left, right, top, and bottom in relation to the fetch.
Also, you can now align it in relation too, aligning it to the top, bottom, left, right, or centered to the fetch.
If the alignment currently selected does not make sense it will default to Center Alignment.